requirements.txt

streamlit==1.25.0
PyPDF2==3.0.1
nltk==3.8.1
numpy==1.24.3
faiss-cpu==1.7.4
torch==2.0.1
sentence-transformers==2.2.2
transformers==4.31.0
